Make characteristics available per working mode (#32)
* feature: include WORKING_MODE in export/import & toolbar logic

The currently selected characteristic must be from the currently selected WORKING_MODE.

* fix: remove unused useEffect

oopsie

* feature: export only characteristics from the currently set working mode

and append the working mode to the file name

* feature: check if user tries to import features from unsupported work modes

(to be changed sooooooooooon)

* fix: fix how we access workingModes from the store

* fix: check if workingMode is not null to set a starting view

* chore: working mode cannot be null (at least it shouldn't be) after setting it for the first time, so we don't have to make it nullable everywhere in the app

:)

* fix: don't replace spaces with underscore while exporting

It will be handled differently when creating new characteristics.

* feature: set working mode to the working mode from the imported file

only if the file specifies a single working mode

* feature: display working mode name in the marking characteristics dialog

as it's confusing for which mode we're setting the characteristics

* fix: don't display characteristics menu entry if working mode is not set

previously workingMode could be an empty string, now it's fixed

* chore: made it mandatory to pass workingMode to getActiveCharacteristicByType

* fix: don't allow keyboard shortcuts to switch to unavailable marking types

If there are no characteristics for the marking type, don't allow the user to switch between them using the keyboard.

* fix: fixed how we retrieve available characteristics per type / work mode

because it caused unexpected bugs when switching work modes.

* fix: fixed an issue where the workingMode was not taken into account when setting the activeCharacteristic array

which would sometimes result in an activeCharacteristic array with silly data in it

* fix: fixed a warning

eslint :3

* chore: make workingMode be stored in tauriStorage, not localStorage

It's better this way, e.g. for consistency across the application.

* fix: oopsie doopsie, missed an import

:x

* chore: remove persistence from the WorkingModeStore

will probably be handled differently

* fix: resolved an issue where importing a single type of characteristic in bulk would not automatically update the working mode

ensuring the user can immediately see what they just imported